Move Linux specific code for Pts to `pts_linux.go`
This allows us to build and test our code on macOS for local
development. With this change, `make test` works on macOs.
